+++ src/main/pi/agent-status-handler-source.ts
@@ -115,7 +115,9 @@ export function getPiAgentStatusHandlerSourceLines(kind: PiAgentKind): string[] 
     '  })',
     '',
     '  // Why: modern Pi stays non-idle across retry/compaction/follow-up work,',
-    '  // while legacy Pi/OMP becomes idle after its final agent_end handlers.',
+    '  // while legacy Pi becomes idle after its final agent_end handlers.',
+    '  // OMP instead marks non-terminal agent_end events with willContinue, so it',
+    '  // returns before the recheck timer is ever armed.',
     '  const AGENT_END_IDLE_RECHECK_MS = 25',
     '  const AGENT_END_IDLE_RECHECK_MAX_MS = 250',
     '  let agentSettledSupported = false',
@@ -169,8 +171,9 @@ export function getPiAgentStatusHandlerSourceLines(kind: PiAgentKind): string[] 
     '',
     "  pi.on('agent_end', (event, ctx) => {",
     ...captureSessionMetadata,
-    '    if (event?.willContinue === true) {',
-    '      clearPendingAgentEndCheck()',
+    '    if (isOmpRuntime()) {',
+    '      if (event?.willContinue === true) return',
+    '      postAgentEndOnce()',
     '      return',
     '    }',
     '    if (agentSettledSupported) return',
